+++ src/components/sections/dynamic/Contact/ContactSection.tsx
@@ -7,10 +7,8 @@ import { Button } from '@/components/elements/Button'
 import { ContactForm } from '@/components/form/ContactForm'
 import { Border } from '@/components/ui/Border'
 import { ContactContent, ContactProps, Office } from '@/types/contact'
-import BasicMarkdown from '@/components/ui/BasicMarkdown'
 import Icon from '@/components/images/Icon'
 import dynamicIconImports from 'lucide-react/dynamicIconImports'
-import { getTranslations } from 'next-intl/server'
 import { useTranslations } from 'next-intl'
 import { Section } from '@/components/ui/Section'
